新增点赞,收藏接口

+++ _internal/article/_internal/event/article/producer.go
@@ -31,6 +31,6 @@ func (p *ProducerKafka) IncrReadCnt(ctx context.Context, event ReadEvent) error 
 }
 
 type ReadEvent struct {
-	Uid int
-	Aid int
+	Uid int64
+	Aid int64
 }
